--- a/scene/gui/dialogs.compat.inc
+++ b/scene/gui/dialogs.compat.inc
@@ -30,6 +30,8 @@
 
 #ifndef DISABLE_DEPRECATED
 
+#include "scene/gui/line_edit.h"
+
 void AcceptDialog::_register_text_enter_bind_compat_89419(Control *p_line_edit) {
 	register_text_enter(Object::cast_to<LineEdit>(p_line_edit));
 }