@dsnopek 2025-03-06
We have some multiple inheritance in the openxr module (which [I've argued for removing](https://github.com/godotengine/godot/pull/101951#issuecomment-2688441647)!), but I'm not exactly sure how it would be affected by this change. Anytime we do `Object::cast_to<T>()` where `T` is an `Object` type, it seems like it'd be fine. Then we could maybe do a couple of `dynamic_cast<T>()`s for the non-`Object` parent class? That would maybe allow the multiple inheritance to keep working. Or, we could stop doing that :-)

> I think this change would break compatibility with GDExtension, as they would now be required to override `derives_from` in `GDCLASS` as well.

Hm, maybe. GDExtension classes are always "attached" to an instance of their nearest parent class that exists within the engine, and any `Object::cast_to<T>()` within the engine would only address classes within the engine, so it might be fine?

In godot-cpp, we'd maybe want to change the way we do `Object::cast_to<T>()` to match the engine, but what we have now (using `dynamic_cast<T>()`) might continue working as-is?

This would require some experimentation.

I could also see a potential affect on one GDExtension using classes from another GDExtension, but that's not a use-case we officially support at the moment anyway - we'd just need to keep this in mind when adding support for it.