@Ivorforce 2025-03-06
To the failing tests, it appears C# objects already declare `derives_from`. That's kinda funny 😄
We'd need to rename either of the functions to make it work again. I'll do that once we have consensus.

> Hm, maybe. GDExtension classes are always "attached" to an instance of their nearest parent class that exists within the engine, and any Object::cast_to<T>() within the engine would only address classes within the engine, so it might be fine?
> In godot-cpp, we'd maybe want to change the way we do Object::cast_to<T>() to match the engine, but what we have now (using dynamic_cast<T>()) might continue working as-is?

Interesting! Yeah, actually with some luck, it would actually be backwards compatible.

Speaking of this, this PR would definitely need to expose this function to GDExtension for overriding. Not sure how to do that.