@Ivorforce 2025-03-06
I searched the issues for `dynamic_cast` and randomly selected https://github.com/godotengine/godot/issues/64398 as a candidate to test real-world impact.
The reproducer profiled 10% impact of `dynamic_cast` for me:
```
1.14 Gc   13.5%	1.14 Gc	 	 Node3D::get_global_transform() const
1.11 Gc   13.1%	1.11 Gc	 	 Node::get_child_count(bool) const
908.59 Mc   10.7%	908.59 Mc	 	 Basis::invert()
892.64 Mc   10.6%	892.64 Mc	 	 __dynamic_cast
715.91 Mc    8.5%	715.91 Mc	 	 Node3DEditorViewport::_calculate_spatial_bounds(Node3D const*, bool, Transform3D const*)
521.04 Mc    6.1%	521.04 Mc	 	 __cxxabiv1::__si_class_type_info::search_below_dst(__cxxabiv1::__dynamic_cast_info*, void const*, int, bool) const
493.45 Mc    5.8%	493.45 Mc	 	 Transform3D::affine_inverse() const
461.81 Mc    5.4%	461.81 Mc	 	 Transform3D::operator*=(Transform3D const&)
447.50 Mc    5.3%	447.50 Mc	 	 AABB::merge_with(AABB const&)
183.30 Mc    2.1%	183.30 Mc	 	 Node::can_process() const
```

On this branch, after exchanging `Object::cast_to`, the entry disappears:
```
1.10 Gc   11.4%	1.10 Gc	 	 Node::get_child_count(bool) const
995.57 Mc   10.3%	995.57 Mc	 	 Node3DEditorViewport::_calculate_spatial_bounds(Node3D const*, bool, Transform3D const*)
796.07 Mc    8.2%	796.07 Mc	 	 Basis::invert()
541.20 Mc    5.6%	541.20 Mc	 	 Node3D::get_global_transform() const
465.41 Mc    4.8%	465.41 Mc	 	 Transform3D::affine_inverse() const
435.89 Mc    4.5%	435.89 Mc	 	 Transform3D::operator*=(Transform3D const&)
344.45 Mc    3.5%	344.45 Mc	 	 AABB::merge_with(AABB const&)
230.18 Mc    2.3%	230.18 Mc	 	 DynamicBVH::optimize_incremental(int)
188.47 Mc    1.9%	188.47 Mc	 	 Node::can_process() const
```

It's a somewhat wonky comparison, because it's not a stress test (although the CPU use is, as the author correctly claimed, unexpectedly high).

But I think it can be concluded that the optimization works as expected in this case.