@lawnjelly 2025-03-09
This is great! Can we test on different compilers ..
can https://godbolt.org/ help here maybe to see the generated code?

I just did a test in a particular intensive scene tree traversal I was doing by just implementing a bitfield in `Node` (as I was only interested in checking for `Spatial` and `VisualInstance`) and this was a boatload faster than `cast_to()`. Increased my overall traversal speed of the `SceneTree` by 1/3 (and I suspect some of the rest might be cache related), so that confirms there is a lot of low hanging fruit here.

(Obviously your approach involves no storage on the node as it works through virtual system, so is better for general use, but this was just a simple bottleneck test.)