@Ivorforce 2025-03-09
> This is great! Maybe we can test on different compilers ..
> can https://godbolt.org/ help here maybe to see the generated code?
>
> Having said that, I'm not sure why a compiler wouldn't inline the inherited calls (at least in release as you say) as they should be able to be evaluated at compile time. I'm pretty sure we must rely on this elsewhere for inheritance macros (I haven't studied them).

I had a look... unexpectedly there's still a ton of optimization potential:
https://godbolt.org/z/b4W3jcqPv

Looks like even retrieving the `typeid` for comparison is complicated.
[Someone suggested](https://stackoverflow.com/a/3865708/730797) comparing pointers first to shortcut the test if the compiler allows, but still doing the 'safe' comparison otherwise.
This all seems unnecessarily complicated to me; it seems like there should be a `typeid` that's just an integer value or something.
This seems out of scope for this PR; while this is not optimal, it's still far better than `dynamic_cast`. We could think about further improvements in a future PR.

> I was doing by just implementing a bitfield in Node

This was my first thought as well, but this would of course only optimize specific classes (if I'm understanding your approach right). 
Still, both could be used if bitfields are needed to optimize specific cases!

> I'd like to also try this in 3.x .. we have no GDExtension there so no binary compatibility to worry about (I'm assuming that's part of the issue with GDExtension?). For recompiled modules I think they would work fine with it.

Agreed.