@Jordyfel 2025-03-17
Reminds me of the discussion here https://github.com/godotengine/godot/pull/80612#issuecomment-1678663349 which mentions that this dynamic cast prevents compiling without RTTI