@Ivorforce 2025-03-17
> Reminds me of the discussion here https://github.com/godotengine/godot/pull/80612#issuecomment-1678663349 which mentions that this dynamic cast prevents compiling without RTTI

This PR still uses RTTI, though it does make a step in a direction of being able to get rid of it.

I also have been conaidering the possibility of getting rid of RTTI too, but it seems Godot makes use of one or two libraries that do use it. I'm not sure how safe it would be to try to enable RTTI just for those. But thats probably a discussion for another time.