@dsnopek 2025-03-27
> We have some multiple inheritance in the openxr module (which [I've argued for removing](https://github.com/godotengine/godot/pull/101951#issuecomment-2688441647)!), but I'm not exactly sure how it would be affected by this change

FYI, we recently got rid of the multiple inheritance in the openxr module in PR https://github.com/godotengine/godot/pull/104087

There's still a `dynamic_cast<T>(..)` in a bit of the deprecated code, but I think we could actually remove that, now that the multiple inheritance is gone. Not super important unless we get to a point where we want to remove RTTI from all the non-thirdparty Godot code, though